backend/requirements.txt
fastapi
uvicorn
asyncpg
databases[postgresql]
python-dotenv